pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trialflow-agro"
version = "0.0.1"
description = "A lightweight, reproducible workflow template for analyzing on-farm agricultural trials."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
  { name = "Gnaneswara (Akhil) Marupilla" }
]

keywords = ["geospatial", "agriculture", "raster", "climate", "workflow"]

dependencies = [
  "numpy",
  "pandas",
  "geopandas",
  "pyyaml",
  "pydantic>=2.0",
  "typer[all]",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "black"
]

[project.urls]
Homepage = "https://github.com/gmarupilla/trialflow-agro"
"Bug Tracker" = "https://github.com/gmarupilla/trialflow-agro/issues"

[tool.setuptools]
packages = ["trialflow_agro"]

[project.scripts]
trialflow-agro = "trialflow_agro.cli.main:main"
